refactor: [#1294] extract server-lib package
