+++ packages/axum-server/Cargo.toml
@@ -4,7 +4,7 @@ description = "A wrapper for the Axum server for Torrust HTTP servers to add tim
 documentation.workspace = true
 edition.workspace = true
 homepage.workspace = true
-keywords = ["axum", "server", "torrust", "torrust", "wrapper"]
+keywords = ["axum", "server", "torrust", "wrapper"]
 license.workspace = true
 name = "torrust-axum-server"
 publish.workspace = true
@@ -21,6 +21,8 @@ hyper = "1"
 hyper-util = { version = "0", features = ["http1", "http2", "tokio"] }
 pin-project-lite = "0"
 tokio = { version = "1", features = ["macros", "net", "rt-multi-thread", "signal", "sync"] }
+torrust-server-lib = { version = "3.0.0-develop", path = "../server-lib" }
 tower = { version = "0", features = ["timeout"] }
+tracing = "0"
 
 [dev-dependencies]
