+++ packages/server-lib/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+authors.workspace = true
+description = "Common functionality used in all Torrust HTTP servers."
+documentation.workspace = true
+edition.workspace = true
+homepage.workspace = true
+keywords = ["lib", "server", "torrust"]
+license.workspace = true
+name = "torrust-server-lib"
+publish.workspace = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[dependencies]
+derive_more = { version = "1", features = ["as_ref", "constructor", "from"] }
+tokio = { version = "1", features = ["macros", "net", "rt-multi-thread", "signal", "sync"] }
+tower-http = { version = "0", features = ["compression-full", "cors", "propagate-header", "request-id", "trace"] }
+tracing = "0"
+
+[dev-dependencies]
