+++ packages/server-lib/README.md
@@ -1,0 +1,11 @@
+# Torrust Server Lib
+
+Common functionality used in all Torrust HTTP servers.
+
+## Documentation
+
+[Crate documentation](https://docs.rs/torrust-axum-server).
+
+## License
+
+The project is licensed under the terms of the [GNU AFFERO GENERAL PUBLIC LICENSE](./LICENSE).
