+++ packages/server-lib/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod logging;
+pub mod registar;
+pub mod signals;
