+++ src/bootstrap/jobs/tracker_apis.rs
@@ -25,6 +25,7 @@ use std::sync::Arc;
 
 use axum_server::tls_rustls::RustlsConfig;
 use tokio::task::JoinHandle;
+use torrust_server_lib::registar::ServiceRegistrationForm;
 use torrust_tracker_configuration::AccessTokens;
 use tracing::instrument;
 
@@ -32,7 +33,6 @@ use super::make_rust_tls;
 use crate::container::HttpApiContainer;
 use crate::servers::apis::server::{ApiServer, Launcher};
 use crate::servers::apis::Version;
-use crate::servers::registar::ServiceRegistrationForm;
 
 /// This is the message that the "launcher" spawned task sends to the main
 /// application process to notify the API server was successfully started.
@@ -97,13 +97,13 @@ async fn start_v1(
 mod tests {
     use std::sync::Arc;
 
+    use torrust_server_lib::registar::Registar;
     use torrust_tracker_test_helpers::configuration::ephemeral_public;
 
     use crate::bootstrap::app::{initialize_app_container, initialize_global_services};
     use crate::bootstrap::jobs::tracker_apis::start_job;
     use crate::container::HttpApiContainer;
     use crate::servers::apis::Version;
-    use crate::servers::registar::Registar;
 
     #[tokio::test]
     async fn it_should_start_http_tracker() {
