+++ src/bootstrap/jobs/udp_tracker.rs
@@ -10,10 +10,10 @@ use std::sync::Arc;
 
 use bittorrent_udp_tracker_core::UDP_TRACKER_LOG_TARGET;
 use tokio::task::JoinHandle;
+use torrust_server_lib::registar::ServiceRegistrationForm;
 use tracing::instrument;
 
 use crate::container::UdpTrackerContainer;
-use crate::servers::registar::ServiceRegistrationForm;
 use crate::servers::udp::server::spawner::Spawner;
 use crate::servers::udp::server::Server;
