+++ src/servers/apis/server.rs
@@ -34,6 +34,10 @@ use futures::future::BoxFuture;
 use thiserror::Error;
 use tokio::sync::oneshot::{Receiver, Sender};
 use torrust_axum_server::custom_axum_server::{self, TimeoutAcceptor};
+use torrust_axum_server::signals::graceful_shutdown;
+use torrust_server_lib::logging::STARTED_ON;
+use torrust_server_lib::registar::{ServiceHealthCheckJob, ServiceRegistration, ServiceRegistrationForm};
+use torrust_server_lib::signals::Halted;
 use torrust_tracker_configuration::AccessTokens;
 use tracing::{instrument, Level};
 
@@ -41,9 +45,6 @@ use super::routes::router;
 use crate::bootstrap::jobs::Started;
 use crate::container::HttpApiContainer;
 use crate::servers::apis::API_LOG_TARGET;
-use crate::servers::logging::STARTED_ON;
-use crate::servers::registar::{ServiceHealthCheckJob, ServiceRegistration, ServiceRegistrationForm};
-use crate::servers::signals::{graceful_shutdown, Halted};
 
 /// Errors that can occur when starting or stopping the API server.
 #[derive(Debug, Error)]
@@ -294,13 +295,13 @@ impl Launcher {
 mod tests {
     use std::sync::Arc;
 
+    use torrust_server_lib::registar::Registar;
     use torrust_tracker_test_helpers::configuration::ephemeral_public;
 
     use crate::bootstrap::app::{initialize_app_container, initialize_global_services};
     use crate::bootstrap::jobs::make_rust_tls;
     use crate::container::HttpApiContainer;
     use crate::servers::apis::server::{ApiServer, Launcher};
-    use crate::servers::registar::Registar;
 
     #[tokio::test]
     async fn it_should_be_able_to_start_and_stop() {
