+++ src/servers/health_check_api/handlers.rs
@@ -2,11 +2,11 @@ use std::collections::VecDeque;
 
 use axum::extract::State;
 use axum::Json;
+use torrust_server_lib::registar::{ServiceHealthCheckJob, ServiceRegistration, ServiceRegistry};
 use tracing::{instrument, Level};
 
 use super::resources::{CheckReport, Report};
 use super::responses;
-use crate::servers::registar::{ServiceHealthCheckJob, ServiceRegistration, ServiceRegistry};
 
 /// Endpoint for container health check.
 ///
