+++ src/servers/udp/server/spawner.rs
@@ -7,11 +7,11 @@ use derive_more::derive::Display;
 use derive_more::Constructor;
 use tokio::sync::oneshot;
 use tokio::task::JoinHandle;
+use torrust_server_lib::signals::Halted;
 
 use super::launcher::Launcher;
 use crate::bootstrap::jobs::Started;
 use crate::container::UdpTrackerContainer;
-use crate::servers::signals::Halted;
 
 #[derive(Constructor, Copy, Clone, Debug, Display)]
 #[display("(with socket): {bind_to}")]
