+++ tests/servers/api/environment.rs
@@ -6,13 +6,13 @@ use bittorrent_tracker_core::authentication::service::AuthenticationService;
 use bittorrent_tracker_core::databases::Database;
 use bittorrent_tracker_core::whitelist::repository::in_memory::InMemoryWhitelist;
 use futures::executor::block_on;
+use torrust_server_lib::registar::Registar;
 use torrust_tracker_api_client::connection_info::{ConnectionInfo, Origin};
 use torrust_tracker_configuration::Configuration;
 use torrust_tracker_lib::bootstrap::app::{initialize_app_container, initialize_global_services};
 use torrust_tracker_lib::bootstrap::jobs::make_rust_tls;
 use torrust_tracker_lib::container::HttpApiContainer;
 use torrust_tracker_lib::servers::apis::server::{ApiServer, Launcher, Running, Stopped};
-use torrust_tracker_lib::servers::registar::Registar;
 use torrust_tracker_primitives::peer;
 
 pub struct Environment<S>
