+++ tests/servers/http/environment.rs
@@ -6,12 +6,12 @@ use bittorrent_tracker_core::databases::Database;
 use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
 use bittorrent_tracker_core::whitelist::manager::WhitelistManager;
 use futures::executor::block_on;
+use torrust_server_lib::registar::Registar;
 use torrust_tracker_configuration::Configuration;
 use torrust_tracker_lib::bootstrap::app::{initialize_app_container, initialize_global_services};
 use torrust_tracker_lib::bootstrap::jobs::make_rust_tls;
 use torrust_tracker_lib::container::HttpTrackerContainer;
 use torrust_tracker_lib::servers::http::server::{HttpServer, Launcher, Running, Stopped};
-use torrust_tracker_lib::servers::registar::Registar;
 use torrust_tracker_primitives::peer;
 
 pub struct Environment<S> {
