+++ tests/servers/udp/environment.rs
@@ -5,10 +5,10 @@ use bittorrent_primitives::info_hash::InfoHash;
 use bittorrent_tracker_core::databases::Database;
 use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
 use bittorrent_udp_tracker_core::statistics;
+use torrust_server_lib::registar::Registar;
 use torrust_tracker_configuration::{Configuration, DEFAULT_TIMEOUT};
 use torrust_tracker_lib::bootstrap::app::{initialize_app_container, initialize_global_services};
 use torrust_tracker_lib::container::UdpTrackerContainer;
-use torrust_tracker_lib::servers::registar::Registar;
 use torrust_tracker_lib::servers::udp::server::spawner::Spawner;
 use torrust_tracker_lib::servers::udp::server::states::{Running, Stopped};
 use torrust_tracker_lib::servers::udp::server::Server;
